feat(KanvazRunner): adds KanvazRunner to run on load
This is quite hacky since I'm only introducing this component in order
to `Query` it's children component. `Query` only allows queries to
Light DOM of that component, which didn't work out for `Kanvaz` itself
since it doesn't have any Light DOM.

This commit adds the `KanvazRunner` component to get hold of all editor
instance to execute the kanvaz on page load.

Also removes `onAllChangesDone` lifecycle and replaces it with `setTimeout()`.
Turned out that `onAllChangesDone` is called infinitely ~15 times per second.